ci: clear pre-existing CI debt — fmt, clippy, rustdoc, audit, fuzz
Every CI check on PR #2 was already failing on main pre-K=17 work;
this commit gets them green so future PRs land with a clean signal.

* `cargo fmt --all` — 107 files reformatted to match the toolchain's
  current rustfmt output (no semantic change).

* `cargo build --all-targets` with `RUSTFLAGS=-D warnings`: 50+ fixes
  across ~30 files. Most are obvious (unused imports, redundant
  closures, manual `is_multiple_of`, `manual_split_once`,
  `useless_format`, `while_let_loop`, `needless_range_loop` where the
  iter rewrite is awkward → `#[allow]` on the test fn, etc.). Two
  scoped crate-level allows:
    - `drift::lib.rs`: `clippy::type_complexity` — 15+ wire-adapter
      `Pin<Box<dyn Future<Output=io::Result<...>>>>` sites are all
      the same SchemeRegistration pattern; refactoring them into
      named aliases is its own project.
    - `drift-ffi::lib.rs`: `clippy::missing_safety_doc` — every
      `unsafe extern "C" fn` shares the same caller-contract,
      documented in drift_ffi.h.

* `RUSTDOCFLAGS=-D warnings cargo doc --no-deps --all-features`:
  13 fixes across drift-git, drift-http, drift-vpn, drift/wire_h2.rs,
  drift/wire_onion.rs. Unwrapped type params (`Vec<u8>`,
  `<user-config>`, etc.) become inline code; bracketed prose
  (`[failover]`, `[vpn]`, `[1]`) becomes inline code; literal `\n`
  in protocol-trace doc comments moved into ```text``` blocks so
  rustdoc doesn't try to parse them as Rust.

* `cargo audit` — add ignore for RUSTSEC-2023-0071 (Marvin Attack
  timing side-channel in the `rsa` crate, reaches us transitively
  via arti-client). No upstream fix is available yet; DRIFT doesn't
  expose RSA decryption to attacker-controlled timing measurements
  directly. Rationale lives in the CI workflow next to the ignore.

* `cargo run --example fuzz_seeds`: cap the federation_directory
  `max_chunk` seed at 10 entries instead of 40. `build_directory`
  debug-asserts at MAX_DIRECTORY_ENTRIES (=10), so the previous
  seed panicked at generation time — which is why every
  fuzz-smoke job was failing. The example also gains the
  PresenceTicket fields that build_directory now requires.

Workspace tests: 108 result blocks pass, 0 fail. No behavior changes.

+++ drift/tests/stream_reliability.rs
@@ -63,7 +63,7 @@ async fn spawn_proxy(target: SocketAddr, profile: LossProfile) -> SocketAddr {
             }
 
             let base = profile.latency_ms;
-            let jitter = rng.gen_range(0..=profile.jitter_ms.saturating_add(1)) as u64;
+            let jitter = rng.gen_range(0..=profile.jitter_ms.saturating_add(1));
             let mut delay = base + jitter;
             if rng.gen::<f64>() < profile.reorder_rate {
                 delay += rng.gen_range(10..=50);
